requirements.txt
# QTCL Blockchain API - Production Dependencies
# Python 3.9+

# Web Framework
Flask==2.3.2
Flask-CORS==4.0.0
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.7

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.0.1

# HTTP & Networking
requests==2.31.0

# Optional: For advanced features
# redis==5.0.0  # For caching/sessions
# Flask-SocketIO==5.3.4  # For WebSocket support
# qiskit==0.43.0  # For quantum computing features
